backend/requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.6
django-cors-headers==4.3.1
djangorestframework==3.15.1
idna==3.7
psycopg==3.1.19
psycopg-binary==3.1.19
python-dateutil==2.9.0.post0
requests==2.31.0
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
termcolor==2.3.0
typing_extensions==4.11.0
urllib3==2.2.1
yaspin==3.0.2
django-filter==24.2
faker==25.8.0
